modifying the login form to use the email and password input of the MUI
