+++ src/components/UI/Molecules/AuthFormFields.tsx
@@ -1,26 +1,13 @@
-import { InputField } from "../Atoms/InputField";
+import { EmailField } from "../Atoms/EmailField";
+import { PasswordField } from "../Atoms/PasswordField";
 
 export const AuthFormFields: React.FC<AuthFormFieldsProps> = ({ formData, handleInputChange }) => {
   return (
     <>
-      <InputField
-        label="Email"
-        name="email"
-        type="email"
-        required
-        placeholder="Email"
-        value={formData.email}
-        onChange={handleInputChange}
-      />
-      <InputField
-        label="Senha"
-        name="password"
-        type="password"
-        required
-        placeholder="Senha"
-        value={formData.password}
-        onChange={handleInputChange}
-      />
+      <EmailField value={formData.email} onChange={handleInputChange} />
+      <div className="mt-3">
+        <PasswordField value={formData.password} onChange={handleInputChange} />
+      </div>
     </>
   );
 };
